fix(infra): align canary alarms and read scopes
