+++ apps/agentcore-canary-runtime/src/config.ts
@@ -2,9 +2,9 @@ import {
 	loadWorkerConfigFromEnv,
 	type WorkerConfig,
 } from "agent-worker/config";
+import { requireEnv } from "agentcore-canary-dispatch/config-utils";
 import {
 	exactSecretArn,
-	requireEnv,
 	verifiedDatabaseUrl,
 } from "agentcore-canary-dispatch/secret-config";
 import { RUNTIME_SHUTDOWN_TIMEOUT_MS } from "./constants";
